Refactor backtest engine responsibilities
