+++ trading/backtest/__init__.py
@@ -1,0 +1,12 @@
+"""Backtest subsystem package."""
+
+from .runner import BacktestRunner
+from .optimizer import BacktestOptimizer
+from .reporting import generate_backtest_charts, generate_recommendations
+
+__all__ = [
+    "BacktestRunner",
+    "BacktestOptimizer",
+    "generate_backtest_charts",
+    "generate_recommendations",
+]
