+++ trading/backtest_engine.py
@@ -7,23 +7,19 @@ ClStock バックテストエンジン
 
 import logging
 from datetime import datetime, timedelta
-from typing import Dict, List, Optional, Tuple, Any
+from typing import Any, Dict, List, Optional, Tuple
 from dataclasses import dataclass, field
-import pandas as pd
 import numpy as np
-import json
-from concurrent.futures import ThreadPoolExecutor, as_completed
-import matplotlib.pyplot as plt
-import seaborn as sns
-from io import BytesIO
-import base64
 
 # 内部モジュール
-from .trading_strategy import TradingStrategy, TradingSignal, SignalType
-from .portfolio_manager import DemoPortfolioManager
-from .risk_manager import DemoRiskManager
-from .trade_recorder import TradeRecorder, PerformanceMetrics
-from .performance_tracker import PerformanceTracker
+from .backtest import (
+    BacktestOptimizer,
+    BacktestRunner,
+    generate_backtest_charts,
+    generate_recommendations,
+)
+from .trade_recorder import PerformanceMetrics
+from .trading_strategy import TradingStrategy
 
 # 既存システム
 from data.stock_data import StockDataProvider
@@ -122,9 +118,6 @@ class BacktestEngine:
         self.trading_strategy.spread_rate = config.spread_rate
         self.trading_strategy.slippage_rate = config.slippage_rate
 
-        # 履歴データキャッシュ
-        self.historical_data: Dict[str, pd.DataFrame] = {}
-
         self.logger = logging.getLogger(__name__)
 
     def run_backtest(
@@ -150,77 +143,10 @@ class BacktestEngine:
             )
             self.logger.info(f"対象銘柄数: {len(target_symbols)}")
 
-            # 履歴データ取得
-            self._load_historical_data(target_symbols)
-
-            # バックテスト実行
-            portfolio_manager = DemoPortfolioManager(self.config.initial_capital)
-            risk_manager = DemoRiskManager(self.config.initial_capital)
-            trade_recorder = TradeRecorder()
-            performance_tracker = PerformanceTracker(self.config.initial_capital)
-
-            # 取引履歴
-            trades_executed = []
-            portfolio_values = []
-            daily_returns = []
-
-            # 日次バックテストループ
-            current_date = self.config.start_date
-            previous_portfolio_value = self.config.initial_capital
-
-            while current_date <= self.config.end_date:
-                try:
-                    # 取引日チェック
-                    if not self._is_trading_day(current_date):
-                        current_date += timedelta(days=1)
-                        continue
-
-                    # 日次取引処理
-                    daily_trades = self._process_trading_day(
-                        current_date,
-                        target_symbols,
-                        portfolio_manager,
-                        risk_manager,
-                        trade_recorder,
-                    )
-
-                    trades_executed.extend(daily_trades)
-
-                    # ポートフォリオ価値更新
-                    portfolio_value = self._calculate_portfolio_value(
-                        current_date, portfolio_manager
-                    )
-                    portfolio_values.append((current_date, portfolio_value))
-
-                    # 日次リターン計算
-                    daily_return = (
-                        portfolio_value - previous_portfolio_value
-                    ) / previous_portfolio_value
-                    daily_returns.append(daily_return)
-                    previous_portfolio_value = portfolio_value
-
-                    # パフォーマンス更新
-                    performance_tracker.update_performance(
-                        portfolio_value,
-                        len(portfolio_manager.positions),
-                        len(daily_trades),
-                    )
-
-                    current_date += timedelta(days=1)
-
-                except Exception as e:
-                    self.logger.error(f"日次処理エラー {current_date}: {e}")
-                    current_date += timedelta(days=1)
-                    continue
-
-            # 結果計算
-            result = self._calculate_backtest_results(
-                trades_executed,
-                portfolio_values,
-                daily_returns,
-                trade_recorder,
-                performance_tracker,
+            runner = BacktestRunner(
+                self.config, self.trading_strategy, self.data_provider, self.logger
             )
+            result = runner.run_backtest(target_symbols)
 
             self.logger.info(f"バックテスト完了: 総リターン {result.total_return:.2%}")
             return result
@@ -341,54 +267,13 @@ class BacktestEngine:
         Returns:
             最適化結果
         """
-        try:
-            best_params = {}
-            best_score = float("-inf")
-            all_results = []
-
-            # パラメータ組み合わせ生成
-            param_combinations = self._generate_parameter_combinations(parameter_ranges)
-
-            self.logger.info(
-                f"パラメータ最適化開始: {len(param_combinations)}通りの組み合わせ"
-            )
-
-            # 並列最適化
-            with ThreadPoolExecutor(max_workers=4) as executor:
-                future_to_params = {
-                    executor.submit(
-                        self._test_parameter_combination, params, optimization_metric
-                    ): params
-                    for params in param_combinations[:50]  # 最初の50通りのみテスト
-                }
-
-                for future in as_completed(future_to_params):
-                    params = future_to_params[future]
-                    try:
-                        score, result = future.result()
-                        all_results.append((params, score, result))
-
-                        if score > best_score:
-                            best_score = score
-                            best_params = params
-
-                        self.logger.info(
-                            f"パラメータテスト完了: {params} スコア: {score:.4f}"
-                        )
-
-                    except Exception as e:
-                        self.logger.error(f"パラメータテストエラー {params}: {e}")
-
-            return {
-                "best_parameters": best_params,
-                "best_score": best_score,
-                "all_results": all_results,
-                "optimization_metric": optimization_metric,
-            }
-
-        except Exception as e:
-            self.logger.error(f"パラメータ最適化エラー: {e}")
-            return {}
+        optimizer = BacktestOptimizer(self.logger)
+        return optimizer.optimize_parameters(
+            base_config=self.config,
+            parameter_ranges=parameter_ranges,
+            optimization_metric=optimization_metric,
+            engine_factory=BacktestEngine,
+        )
 
     def generate_backtest_report(self, result: BacktestResult) -> Dict[str, Any]:
         """バックテストレポート生成"""
@@ -437,7 +322,7 @@ class BacktestEngine:
             trade_analysis = self._analyze_trades(result.trade_history)
 
             # チャート生成
-            charts = self._generate_backtest_charts(result)
+            charts = generate_backtest_charts(result, logger=self.logger)
 
             return {
                 "basic_statistics": basic_stats,
@@ -446,7 +331,7 @@ class BacktestEngine:
                 "monthly_analysis": monthly_analysis,
                 "trade_analysis": trade_analysis,
                 "charts": charts,
-                "recommendations": self._generate_recommendations(result),
+                "recommendations": generate_recommendations(result),
             }
 
         except Exception as e:
@@ -454,469 +339,6 @@ class BacktestEngine:
             return {}
 
     # --- プライベートメソッド ---
-
-    def _load_historical_data(self, symbols: List[str]):
-        """履歴データ読み込み"""
-        self.logger.info("履歴データ読み込み開始")
-
-        for symbol in symbols:
-            try:
-                # 期間を少し拡張（テクニカル指標計算用）
-                extended_start = self.config.start_date - timedelta(days=100)
-
-                data = self.data_provider.get_stock_data(
-                    symbol, start_date=extended_start, end_date=self.config.end_date
-                )
-
-                if data is not None and len(data) > 0:
-                    # テクニカル指標計算
-                    data = self.data_provider.calculate_technical_indicators(data)
-                    self.historical_data[symbol] = data
-                    self.logger.info(
-                        f"履歴データ読み込み完了: {symbol} ({len(data)}日分)"
-                    )
-                else:
-                    self.logger.warning(f"履歴データ取得失敗: {symbol}")
-
-            except Exception as e:
-                self.logger.error(f"履歴データ読み込みエラー {symbol}: {e}")
-
-    def _process_trading_day(
-        self,
-        date: datetime,
-        symbols: List[str],
-        portfolio_manager: DemoPortfolioManager,
-        risk_manager: DemoRiskManager,
-        trade_recorder: TradeRecorder,
-    ) -> List[Dict[str, Any]]:
-        """取引日処理"""
-        trades_executed = []
-
-        try:
-            # ポジション更新
-            portfolio_manager.update_positions()
-
-            # 各銘柄の取引機会チェック
-            for symbol in symbols:
-                if symbol not in self.historical_data:
-                    continue
-
-                # 指定日までのデータ取得
-                historical_data = self.historical_data[symbol]
-                date_mask = historical_data.index <= date
-                available_data = historical_data[date_mask]
-
-                if len(available_data) < 50:  # 最小データ要件
-                    continue
-
-                # 取引シグナル生成（バックテスト用）
-                signal = self._generate_backtest_signal(symbol, available_data, date)
-
-                if signal is None:
-                    continue
-
-                # リスク管理チェック
-                current_capital = portfolio_manager.current_cash + sum(
-                    pos.market_value for pos in portfolio_manager.positions.values()
-                )
-
-                if not risk_manager.can_open_position(symbol, signal.position_size):
-                    continue
-
-                # バックテスト取引実行
-                trade = self._execute_backtest_trade(
-                    signal, date, portfolio_manager, trade_recorder
-                )
-
-                if trade:
-                    trades_executed.append(trade)
-
-        except Exception as e:
-            self.logger.error(f"取引日処理エラー {date}: {e}")
-
-        return trades_executed
-
-    def _generate_backtest_signal(
-        self, symbol: str, historical_data: pd.DataFrame, current_date: datetime
-    ) -> Optional[TradingSignal]:
-        """バックテスト用シグナル生成"""
-        try:
-            # 87%精度システムでシグナル生成（履歴データベース）
-            # 現在日時点でのデータのみ使用
-            current_price = historical_data.loc[
-                historical_data.index <= current_date, "Close"
-            ].iloc[-1]
-
-            # 簡略化された予測（実際には87%システムを使用）
-            # ここでは過去データに基づく簡易予測を実装
-            signal_data = self._calculate_historical_signal(
-                symbol, historical_data, current_date
-            )
-
-            if signal_data["confidence"] < self.config.confidence_threshold:
-                return None
-
-            return TradingSignal(
-                symbol=symbol,
-                signal_type=signal_data["signal_type"],
-                confidence=signal_data["confidence"],
-                predicted_price=signal_data["predicted_price"],
-                current_price=current_price,
-                expected_return=signal_data["expected_return"],
-                position_size=signal_data["position_size"],
-                timestamp=current_date,
-                reasoning=signal_data["reasoning"],
-                precision_87_achieved=signal_data["precision_87_achieved"],
-            )
-
-        except Exception as e:
-            self.logger.error(f"バックテストシグナル生成エラー {symbol}: {e}")
-            return None
-
-    def _calculate_historical_signal(
-        self, symbol: str, data: pd.DataFrame, date: datetime
-    ) -> Dict[str, Any]:
-        """履歴データベース簡易シグナル計算"""
-        try:
-            # 指定日までのデータ
-            current_data = data[data.index <= date]
-
-            if len(current_data) < 20:
-                return self._default_signal()
-
-            # 現在価格
-            current_price = current_data["Close"].iloc[-1]
-
-            # 簡易テクニカル分析
-            sma_20 = current_data["Close"].rolling(20).mean().iloc[-1]
-            rsi = self._calculate_rsi(current_data["Close"], 14).iloc[-1]
-
-            # トレンド分析
-            price_change = (
-                current_price - current_data["Close"].iloc[-20]
-            ) / current_data["Close"].iloc[-20]
-
-            # シグナル判定
-            signal_strength = 0.0
-            reasoning_parts = []
-
-            # RSIベースシグナル
-            if rsi < 30:
-                signal_strength += 0.3
-                reasoning_parts.append("RSI過売り")
-            elif rsi > 70:
-                signal_strength -= 0.3
-                reasoning_parts.append("RSI過買い")
-
-            # トレンドベースシグナル
-            if current_price > sma_20:
-                signal_strength += 0.4
-                reasoning_parts.append("SMA20上回り")
-            else:
-                signal_strength -= 0.4
-                reasoning_parts.append("SMA20下回り")
-
-            # ボラティリティ調整
-            volatility = current_data["Close"].pct_change().std()
-            confidence = min(abs(signal_strength) + 0.3, 0.9)
-
-            # 87%精度達成判定（ランダム）
-            precision_87_achieved = np.random.random() < 0.3  # 30%の確率
-
-            # 期待リターン計算
-            expected_return = signal_strength * 0.05  # 最大5%
-
-            # ポジションサイズ計算
-            position_size = (
-                self.config.initial_capital * self.config.max_position_size * confidence
-            )
-
-            # シグナルタイプ決定
-            if signal_strength > 0.3:
-                signal_type = SignalType.BUY
-            elif signal_strength < -0.3:
-                signal_type = SignalType.SELL
-            else:
-                signal_type = SignalType.HOLD
-
-            # 予測価格
-            predicted_price = current_price * (1 + expected_return)
-
-            return {
-                "signal_type": signal_type,
-                "confidence": confidence,
-                "predicted_price": predicted_price,
-                "expected_return": expected_return,
-                "position_size": position_size,
-                "reasoning": " | ".join(reasoning_parts),
-                "precision_87_achieved": precision_87_achieved,
-            }
-
-        except Exception as e:
-            self.logger.error(f"履歴シグナル計算エラー: {e}")
-            return self._default_signal()
-
-    def _calculate_rsi(self, prices: pd.Series, window: int = 14) -> pd.Series:
-        """RSI計算"""
-        delta = prices.diff()
-        gain = (delta.where(delta > 0, 0)).rolling(window=window).mean()
-        loss = (-delta.where(delta < 0, 0)).rolling(window=window).mean()
-        rs = gain / loss
-        return 100 - (100 / (1 + rs))
-
-    def _default_signal(self) -> Dict[str, Any]:
-        """デフォルトシグナル"""
-        return {
-            "signal_type": SignalType.HOLD,
-            "confidence": 0.0,
-            "predicted_price": 0.0,
-            "expected_return": 0.0,
-            "position_size": 0.0,
-            "reasoning": "データ不足",
-            "precision_87_achieved": False,
-        }
-
-    def _execute_backtest_trade(
-        self,
-        signal: TradingSignal,
-        date: datetime,
-        portfolio_manager: DemoPortfolioManager,
-        trade_recorder: TradeRecorder,
-    ) -> Optional[Dict[str, Any]]:
-        """バックテスト取引実行"""
-        try:
-            # スリッページ適用
-            slippage = np.random.normal(0, self.config.slippage_rate)
-            execution_price = signal.current_price * (1 + slippage)
-
-            # 数量計算
-            quantity = int(signal.position_size / execution_price)
-            if quantity <= 0:
-                return None
-
-            actual_position_value = quantity * execution_price
-
-            # 取引コスト計算
-            commission = actual_position_value * self.config.commission_rate
-            spread_cost = actual_position_value * self.config.spread_rate
-            total_cost = commission + spread_cost
-
-            # ポートフォリオ更新
-            portfolio_manager.add_position(
-                signal.symbol, quantity, execution_price, signal.signal_type
-            )
-
-            # 取引記録
-            trade_data = {
-                "trade_id": f"{signal.symbol}_{date.strftime('%Y%m%d')}",
-                "symbol": signal.symbol,
-                "action": "OPEN",
-                "quantity": quantity,
-                "price": execution_price,
-                "timestamp": date.isoformat(),
-                "signal_type": signal.signal_type.value,
-                "confidence": signal.confidence,
-                "precision_87_achieved": signal.precision_87_achieved,
-                "expected_return": signal.expected_return,
-                "position_size": actual_position_value,
-                "trading_costs": {
-                    "commission": commission,
-                    "spread": spread_cost,
-                    "total_cost": total_cost,
-                },
-            }
-
-            trade_recorder.record_trade(trade_data)
-
-            return trade_data
-
-        except Exception as e:
-            self.logger.error(f"バックテスト取引実行エラー: {e}")
-            return None
-
-    def _calculate_portfolio_value(
-        self, date: datetime, portfolio_manager: DemoPortfolioManager
-    ) -> float:
-        """ポートフォリオ価値計算"""
-        try:
-            total_value = portfolio_manager.current_cash
-
-            for symbol, position in portfolio_manager.positions.items():
-                if symbol in self.historical_data:
-                    # 指定日の価格取得
-                    historical_data = self.historical_data[symbol]
-                    date_mask = historical_data.index <= date
-                    available_data = historical_data[date_mask]
-
-                    if len(available_data) > 0:
-                        current_price = available_data["Close"].iloc[-1]
-                        position_value = position.quantity * current_price
-                        total_value += position_value
-
-            return total_value
-
-        except Exception as e:
-            self.logger.error(f"ポートフォリオ価値計算エラー: {e}")
-            return portfolio_manager.current_cash
-
-    def _calculate_backtest_results(
-        self,
-        trades: List[Dict[str, Any]],
-        portfolio_values: List[Tuple[datetime, float]],
-        daily_returns: List[float],
-        trade_recorder: TradeRecorder,
-        performance_tracker: PerformanceTracker,
-    ) -> BacktestResult:
-        """バックテスト結果計算"""
-        try:
-            if not portfolio_values:
-                return self._empty_backtest_result()
-
-            # 基本統計
-            initial_value = self.config.initial_capital
-            final_value = portfolio_values[-1][1]
-            total_return = (final_value - initial_value) / initial_value
-
-            # 期間計算
-            days = (self.config.end_date - self.config.start_date).days
-            annualized_return = (
-                (1 + total_return) ** (252 / days) - 1 if days > 0 else 0
-            )
-
-            # リスク指標
-            volatility = np.std(daily_returns) * np.sqrt(252) if daily_returns else 0
-            sharpe_ratio = annualized_return / volatility if volatility > 0 else 0
-
-            # ダウンサイドリスク
-            downside_returns = [r for r in daily_returns if r < 0]
-            downside_volatility = (
-                np.std(downside_returns) * np.sqrt(252) if downside_returns else 0
-            )
-            sortino_ratio = (
-                annualized_return / downside_volatility
-                if downside_volatility > 0
-                else 0
-            )
-
-            # 最大ドローダウン
-            max_drawdown = self._calculate_max_drawdown(portfolio_values)
-            calmar_ratio = (
-                annualized_return / abs(max_drawdown) if max_drawdown != 0 else 0
-            )
-
-            # 取引統計
-            completed_trades = [t for t in trades if "profit_loss" in t]
-            total_trades = len(completed_trades)
-            winning_trades = len(
-                [t for t in completed_trades if t.get("profit_loss", 0) > 0]
-            )
-            win_rate = winning_trades / total_trades * 100 if total_trades > 0 else 0
-
-            # プロフィットファクター
-            profits = [
-                t["profit_loss"]
-                for t in completed_trades
-                if t.get("profit_loss", 0) > 0
-            ]
-            losses = [
-                abs(t["profit_loss"])
-                for t in completed_trades
-                if t.get("profit_loss", 0) < 0
-            ]
-            total_profits = sum(profits) if profits else 0
-            total_losses = sum(losses) if losses else 0
-            profit_factor = (
-                total_profits / total_losses if total_losses > 0 else float("inf")
-            )
-
-            # 87%精度統計
-            precision_87_trades = len(
-                [t for t in trades if t.get("precision_87_achieved", False)]
-            )
-            precision_87_wins = len(
-                [
-                    t
-                    for t in completed_trades
-                    if t.get("precision_87_achieved", False)
-                    and t.get("profit_loss", 0) > 0
-                ]
-            )
-            precision_87_success_rate = (
-                precision_87_wins / precision_87_trades * 100
-                if precision_87_trades > 0
-                else 0
-            )
-
-            # VaR・期待ショートフォール
-            var_95 = np.percentile(daily_returns, 5) if daily_returns else 0
-            tail_returns = [r for r in daily_returns if r <= var_95]
-            expected_shortfall = np.mean(tail_returns) if tail_returns else var_95
-
-            # コスト計算
-            total_costs = sum(
-                t.get("trading_costs", {}).get("total_cost", 0) for t in trades
-            )
-            total_tax = total_profits * self.config.tax_rate
-
-            # ベンチマーク比較（簡略化）
-            benchmark_return = 0.05  # 仮の年率5%
-            excess_return = annualized_return - benchmark_return
-            beta = 1.0  # 簡略化
-            alpha = excess_return
-
-            return BacktestResult(
-                config=self.config,
-                start_date=self.config.start_date,
-                end_date=self.config.end_date,
-                total_return=total_return,
-                annualized_return=annualized_return,
-                volatility=volatility,
-                sharpe_ratio=sharpe_ratio,
-                sortino_ratio=sortino_ratio,
-                max_drawdown=max_drawdown,
-                calmar_ratio=calmar_ratio,
-                total_trades=total_trades,
-                win_rate=win_rate,
-                profit_factor=profit_factor,
-                precision_87_trades=precision_87_trades,
-                precision_87_success_rate=precision_87_success_rate,
-                final_value=final_value,
-                benchmark_return=benchmark_return,
-                excess_return=excess_return,
-                beta=beta,
-                alpha=alpha,
-                information_ratio=excess_return / volatility if volatility > 0 else 0,
-                var_95=var_95,
-                expected_shortfall=expected_shortfall,
-                total_costs=total_costs,
-                total_tax=total_tax,
-                daily_returns=daily_returns,
-                trade_history=trades,
-                portfolio_values=portfolio_values,
-            )
-
-        except Exception as e:
-            self.logger.error(f"バックテスト結果計算エラー: {e}")
-            return self._empty_backtest_result()
-
-    def _calculate_max_drawdown(
-        self, portfolio_values: List[Tuple[datetime, float]]
-    ) -> float:
-        """最大ドローダウン計算"""
-        if not portfolio_values:
-            return 0.0
-
-        values = [v[1] for v in portfolio_values]
-        peak = values[0]
-        max_drawdown = 0.0
-
-        for value in values:
-            if value > peak:
-                peak = value
-            drawdown = (peak - value) / peak if peak > 0 else 0
-            max_drawdown = max(max_drawdown, drawdown)
-
-        return max_drawdown
 
     def _is_trading_day(self, date: datetime) -> bool:
         """取引日判定"""
@@ -971,49 +393,6 @@ class BacktestEngine:
             portfolio_values=[],
         )
 
-    def _generate_parameter_combinations(
-        self, parameter_ranges: Dict[str, List[float]]
-    ) -> List[Dict[str, float]]:
-        """パラメータ組み合わせ生成"""
-        import itertools
-
-        keys = list(parameter_ranges.keys())
-        values = list(parameter_ranges.values())
-
-        combinations = []
-        for combo in itertools.product(*values):
-            combinations.append(dict(zip(keys, combo)))
-
-        return combinations
-
-    def _test_parameter_combination(
-        self, params: Dict[str, float], metric: str
-    ) -> Tuple[float, BacktestResult]:
-        """パラメータ組み合わせテスト"""
-        # 設定更新
-        test_config = BacktestConfig(
-            start_date=self.config.start_date,
-            end_date=self.config.end_date,
-            initial_capital=self.config.initial_capital,
-            precision_threshold=params.get(
-                "precision_threshold", self.config.precision_threshold
-            ),
-            confidence_threshold=params.get(
-                "confidence_threshold", self.config.confidence_threshold
-            ),
-            max_position_size=params.get(
-                "max_position_size", self.config.max_position_size
-            ),
-            target_symbols=self.config.target_symbols,
-        )
-
-        # テストバックテスト実行
-        test_engine = BacktestEngine(test_config)
-        result = test_engine.run_backtest()
-
-        # メトリクス取得
-        score = getattr(result, metric, 0.0)
-        return score, result
 
     def _calculate_combined_performance(
         self, period_results: List[BacktestResult]
@@ -1168,69 +547,3 @@ class BacktestEngine:
             ),
             "avg_holding_period": 1.0,  # 簡略化
         }
-
-    def _generate_backtest_charts(self, result: BacktestResult) -> Dict[str, str]:
-        """バックテストチャート生成"""
-        charts = {}
-
-        try:
-            # 資産曲線
-            if result.portfolio_values:
-                fig, ax = plt.subplots(figsize=(12, 6))
-                dates = [pv[0] for pv in result.portfolio_values]
-                values = [pv[1] for pv in result.portfolio_values]
-
-                ax.plot(dates, values, linewidth=2, label="ポートフォリオ価値")
-                ax.axhline(
-                    y=result.config.initial_capital,
-                    color="red",
-                    linestyle="--",
-                    alpha=0.5,
-                    label="初期資本",
-                )
-                ax.set_title("バックテスト資産曲線", fontweight="bold")
-                ax.set_ylabel("価値 (円)")
-                ax.legend()
-                ax.grid(True, alpha=0.3)
-                plt.xticks(rotation=45)
-                plt.tight_layout()
-
-                charts["equity_curve"] = self._fig_to_base64(fig)
-
-        except Exception as e:
-            self.logger.error(f"チャート生成エラー: {e}")
-
-        return charts
-
-    def _generate_recommendations(self, result: BacktestResult) -> List[str]:
-        """推奨事項生成"""
-        recommendations = []
-
-        if result.sharpe_ratio < 1.0:
-            recommendations.append("シャープレシオが低いため、リスク調整が必要です")
-
-        if result.max_drawdown > 0.2:
-            recommendations.append(
-                "最大ドローダウンが大きいため、リスク管理を強化してください"
-            )
-
-        if result.win_rate < 50:
-            recommendations.append(
-                "勝率が低いため、エントリー条件の見直しを検討してください"
-            )
-
-        if result.precision_87_success_rate < 60:
-            recommendations.append(
-                "87%精度取引の成功率が低いため、精度判定基準の調整が必要です"
-            )
-
-        return recommendations
-
-    def _fig_to_base64(self, fig) -> str:
-        """図をBase64エンコード"""
-        buffer = BytesIO()
-        fig.savefig(buffer, format="png", dpi=100, bbox_inches="tight")
-        buffer.seek(0)
-        image_base64 = base64.b64encode(buffer.getvalue()).decode()
-        plt.close(fig)
-        return image_base64
